requirements.txt
pyld
requests
pyparsing
pyshacl
python-decouple
rdflib
